requirements.txt
kafka-python==2.0.2
requests==2.31.0
pandas==2.2.2
scikit-learn==1.3.0
sqlalchemy==2.0.18
psycopg2-binary==2.9.7
python-dateutil==2.8.2
